Version 0.2.12 - Fixed group scaling issue
On the end of group scale, the scale factor is set back to 1 and
the width/height and stroke width of the relevant shapes (not
free draw paths) are set to the percieved values. (Mostly just
`width * scale` etc).

Addresses https://www.pivotaltracker.com/story/show/74634376

Also took container styling out of the drawtool styling as per
[Piotr's suggestion](https://github.com/concord-consortium/drawing-tool/commit/16d5aaa1001a92c4b595e7c065a56d76efc474b2#commitcomment-6951794)