Version 0.4.1 - Fixed line rescaling issue, updated version number
The last commit (cda3536) also fixed the bug in question (https://www.pivotaltracker.com/story/show/74641772).

Note: we need to change the control points to better center them on the line endpoints. Right now, if we make a line width too large, the offset is visible and ugly. I made a note in the code (as of right now @line-tool.js ln 123)