getActiveGroup -> getActiveObjects
This replaces all calls of getActiveGroup with getActiveObjects,
which returns an array of objects.

If there is only one active object, it returns an array of length 1,
which slightly simplifies some of the getActiveObject/getActiveGroup
if-then statements.

In the one case where we keep getActiveObject, we change the
type-checking to activeSelection.