Added option to specify buttons shown in UI
This allows for both naming the buttons shown and setting their order.

This is the first of a few stories for the 2.1.0 release so it has been versioned as a pre-release in the package.json file.

As a cleanup I removed the yarn.lock and references to yarn as the package-lock.json file was newer and the yarn.lock hadn't been touched in 4 years.