fix: Text, stroke and fill palettes not showing with custom buttons [PT-182016906]
This adds code to also load the palette buttons when the text, stroke and fill tools are used in the custom button settings.

NOTE: Although this is a bugfix there was already a 2.3.0-pre.1 package on npm so I had to use a minor version upgrade instead of a bugfix semver bump.